# frontend/web/src/api/openapi.json
{
 "components": {
  "schemas": {
...
   "Comp": {
    "properties": {
     "actor": {
      "additionalProperties": true,
      "title": "Actor",
      "type": "object"
     },
     "email": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Email"
     },
     "expires_at": {
      "format": "date-time",
      "title": "Expires At",
      "type": "string"
     },
     "id": {
      "format": "uuid",
      "title": "Id",
      "type": "string"
     },
     "reader_id": {
      "anyOf": [
       {
        "format": "uuid",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reader Id"
     },
     "reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reason"
     },
     "revoke_reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Revoke Reason"
     },
     "revoked_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Revoked At"
     },
     "revoked_by": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Revoked By"
     },
     "running": {
      "title": "Running",
      "type": "boolean"
     },
     "source": {
      "title": "Source",
      "type": "string"
     },
     "started_at": {
      "format": "date-time",
      "title": "Started At",
      "type": "string"
     }
    },
    "required": [
     "id",
     "reader_id",
     "email",
     "source",
     "started_at",
     "expires_at",
     "reason",
     "actor",
     "revoked_at",
     "revoked_by",
     "revoke_reason",
     "running"
    ],
    "title": "Comp",
    "type": "object"
   },
...
   "GrantComp": {
    "properties": {
     "company": {
      "anyOf": [
       {
        "maxLength": 100,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Company"
     },
     "email": {
      "maxLength": 254,
      "title": "Email",
      "type": "string"
     },
     "reason": {
      "maxLength": 500,
      "minLength": 1,
      "title": "Reason",
      "type": "string"
     },
     "until": {
      "format": "date-time",
      "title": "Until",
      "type": "string"
     }
    },
    "required": [
     "email",
     "until",
     "reason"
    ],
    "title": "GrantComp",
    "type": "object"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "Revoke": {
    "properties": {
     "reason": {
      "maxLength": 500,
      "minLength": 1,
      "title": "Reason",
      "type": "string"
     }
    },
    "required": [
     "reason"
    ],
    "title": "Revoke",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "scheme": "bearer",
    "type": "http"
   }
  }
 },
 "info": {
  "title": "Autora API",
  "version": "0.0.1"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/admin/memberships/comps": {
   "get": {
    "operationId": "list_comps_api_admin_memberships_comps_get",
    "parameters": [
     {
      "in": "query",
      "name": "company",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "maxLength": 100,
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Company"
      }
     },
     {
      "in": "query",
      "name": "running",
      "required": false,
      "schema": {
       "default": false,
       "title": "Running",
       "type": "boolean"
      }
     },
     {
      "in": "cookie",
      "name": "autora_admin",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Autora Admin"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/Comp"
         },
         "title": "Response List Comps Api Admin Memberships Comps Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "List Comps",
    "tags": [
     "admin-memberships"
    ]
   },
   "post": {
    "description": "Give a reader VIP until ``until``. No order, no payment, no revenue.",
    "operationId": "grant_comp_api_admin_memberships_comps_post",
    "parameters": [
     {
      "in": "cookie",
      "name": "autora_admin",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Autora Admin"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/GrantComp"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Comp"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Grant Comp",
    "tags": [
     "admin-memberships"
    ]
   }
  },
  "/api/admin/memberships/comps/{grant_id}": {
   "get": {
    "operationId": "get_comp_api_admin_memberships_comps__grant_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "grant_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Grant Id",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "company",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "maxLength": 100,
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Company"
      }
     },
     {
      "in": "cookie",
      "name": "autora_admin",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Autora Admin"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Comp"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Get Comp",
    "tags": [
     "admin-memberships"
    ]
   }
  },
  "/api/admin/memberships/comps/{grant_id}/revoke": {
   "post": {
    "description": "End a comp now. The row is kept, with who ended it, when and why.",
    "operationId": "revoke_comp_api_admin_memberships_comps__grant_id__revoke_post",
    "parameters": [
     {
      "in": "path",
      "name": "grant_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Grant Id",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "company",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "maxLength": 100,
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Company"
      }
     },
     {
      "in": "cookie",
      "name": "autora_admin",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Autora Admin"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/Revoke"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Comp"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Revoke Comp",
    "tags": [
     "admin-memberships"
    ]
   }
  },
...
 }
}